c1sub

+++ math_operations_h/operations.h
@@ -8,7 +8,7 @@ extern int OPR_iSoma(int iNum1, int iNum2, int iNum3);
 
 extern int OPR_iProduto(int iNum1, int iNum2);
 
-extern int OPR_iPotencia(int iNum1, int iExp);
+extern long OPR_iPotencia(long iNum1, long iExp);
 
 extern long OPR_lFatorial(long lNumber);
